apps/daemon/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv
GitPython
mss
google-genai
pillow
